LayoutTests/imported/w3c/web-platform-tests/custom-elements/form-associated/label-delegatesFocus-expected.txt

label

FAIL Clicking on a label for a form associated custom element with delegatesFocus should focus the custom element's focus delegate. assert_equals: expected Element node <my-custom-element id="custom"></my-custom-element> but got Element node <body><form>
  <label for="custom">label</label>
  <my-cu...
